LU-5292 lfsck: skip orphan handling only for failed servers
The layout LFSCK will record the failed servers in the LFSCK tracing
file (lfsck_layout) during the first-stage scanning, then when moves
to the second-stage scanning, the LFSCK will know which OSTs contain
the OST-objects that have not been verified or failed to be verified
during the first-stage scanning. Then the LFSCK will skip the orphan
OST-objects handling on those OSTs. But other OSTs can be handled as
normal case without affected by the failed OSTs.

This patch also builds the framework of recording failed servers for
namespace LFSCK, which will be done in subsequent patches.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I4d7a9fc2e22cb8c6ef1c4cf73383ec588c95da53