LU-4906 llite: read page from LMV/MDC for readdir
During rm -rf, if statahead is enable, it might meet ldlm lock
ping pong, because the statahead thread will enqueue the lock
for each entry in mdc_read_entry, in the mean time, MDT will
revoke the lock when it deletes each entry, which caused a lot
extra RPC.

To avoid such ping pong behavior for ldlm lock, read dir will
read page, instead of entry, from MDC each time. Note: after it
get the entry page, it will not hold the lock, but only hold
the page reference. So if the lock was canceled, and the page
might become stale, but it only causes the following stat ahead
thread to try to getattr stale entries from MDT, which is fine
for stat-ahead process.

Signed-off-by: wang di <di.wang@intel.com>
Change-Id: I189f0039b5624455b70d1381b43d452cc8766543