LU-5134 utils: Add parallel option to lctl set_param
Add a "-p" option to lctl set_param to enable setting multiple matched
parameters in parallel.  When called with "-p", lctl will set up a
work queue of matched file names and spawn a fixed number of threads
per CPU.  Each thread will pull items off the work queue, write to the
file associated with each work item, and return when there are no more
items on the work queue.

A bit field called po_parallel is added to the struct param_opts to
indicate whether the set_param should be executed in parallel.  If in
parallel, jt_lcfg_setparam calls setparam_display_parallel, which sets
up a work queue and spawns setparam_threads.  Otherwise, it calls
setparam_display, which iterates through the matched files and sets
them in series.  In each case, the common work of writing the value to
each file is handled by the function set_one_param().

If lctl is compiled without pthread support, "lctl set_param" will
still accept the "-p" option, but it will print an error message, and
it will set the parameters in series.

Sanity test 124c was added to test the "-p" option to lctl set_param.
It also checks that writing "clear" to lru_size clears all unused
locks.  There was no test for this before.

Signed-off-by: Ryan Haasken <haasken@cray.com>
Change-Id: I3f96a6f06c50d4ba2ce97050c35f46b976dfc005