LU-4856 misc: 64 bit counters for objects tied to memory size
When the number of an object in use or circulation is tied to memory
size of the system, very large memory systems can overflow 32 bit
counters.  This patch converts several of the most obviously exposed
counters - those initialized on the basis of memory size.
This incudes the debug log size, client osc pool/lru, and ldlm pool.
Several additional checks are made to detect overflow on variables
which have not been converted.

Signed-off-by: Stephen Champion <schamp@sgi.com>
Change-Id: I8deff55d7cc9774722f3dc38a2c7b15877e698f0