LU-2544 lnet: SMP scaling api for Gemini driver
Increase the performance of the Gemini LND driver by
using the Lustre SMP api. Bind the threads to specific
cores to limit the cost of thread migration between
different cores. The second improvement is allocating
memory based on which core so we can keep the cache
warm to limit cache line misses.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I0c1fca8d6ea21986cddfb615eb455ae09843a4dc