LU-3382 build: clean unused link dependencies
Compiled executable Lustre files such as:
lst, lstclient, loadgen, lustre_rsync, lctl, lfs,
debugctl and ptlctl are linked against libtinfo.so.5
and libncurses.so.5, however none of the library's
symbols are used. Same issue with multiop which is
linked against librt.so.1, however uses none of the
library's symbols, and liblustre.so which is linked
against libz.so.1 and uses none of the library's
symbols as well.

Signed-off-by: Thomas Stibor <thomas@stibor.net>
Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I861878ca93dc0a07e2d4d718870a960eb86974a9