LU-4800 llite: handle lustre fs registration properly for newer kernels
With newer kernels running the mount command no longers properly loads
the kernel modules needed so a client mount fails. This patch addresses
this by loading the module lustre in the obdclass module when the routine
lustre_fill_super is executed. We also avoid the module cycular dependency
between these two modules by using symbol_get to retrieve the functions
from llite to be executed in the obdclass context.

Change-Id: I4d091bb8ba94be07a01146031b1c097565ba113e
Signed-off-by: James Simmons <uja.ornl@gmail.com>