LU-1778 libcfs: add a service that prints a nidlist
The libcfs already provides services to parse a string into a nidlist
and to match a nid into a nidlist. This patch implements a service
that prints a nidlist into a buffer.

This is required for instance to print the nosquash_nids parameter
of the MDT procfs component.

Additionally, this patch fixes a bug in return code of
parse_addrange() routine, so that parsing of nids including
a * character works fine ('*@elan' for instance).

Lustre-change: http://review.whamcloud.com/9221
Lustre-commit: 917b0a17d4e50db9f966763925a2dc860d932a3b

Signed-off-by: Gregoire Pichon <gregoire.pichon@bull.net>
Change-Id: Ifd5a9c224d43c815104eb170c5b854a1951ab47a