LU-2675 libcfs: remove unused open flag conversions
Remove the unused cfs_oflag2univ() and cfs_univ2oflag() conversion
routines.  These were intended to convert O_* open flags to and from
platform-specific flags at the user interface, but are not used
anywhere in the code.  The conversion of the client-specific open
flags to the Lustre protocol flags is done in mds_pack_open_flags()
and there is no need to ever convert back to client-specific flags.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I1f5f7fb7eefa2f8dff161f02b37bee6a9d500c1e