LU-5252: Do not unconditionally delete the HSM event fifo
The lustreapi HSM library will use a pipe if it already exists.
However the deregistration will unconditionally delete that pipe,
which is not correct.

An admin may want to create the pipe with certain rights/ownership,
which the llapi_hsm_register_event_fifo() doesn't allow. In that case,
llapi_hsm_unregister_event_fifo() should not delete it.

Change-Id: I487f6efea06c6f6bc88b2f838153c0239e369461
Signed-off-by: frank zago <fzago@cray.com>