LU-4239 mdt: Remove pli_fids & pli_fidcount
This bug occurs because all of the fids found during
a path lookup were being stored in pli_fids and the
array was too small for very deep paths.

Increasing the array size to an appropriate value takes
~32K of memory. (PS1 does this.)

Currently, only one of the stored FIDs is used.  This patch
removes the array entirely and stores the relevant fid
in pli_fid instead.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I357b4d91b9c65a836d289d5815c763306ac34565