LU-4788 lfsck: namespace LFSCK handles remote object (1)
During the first-stage scanning, if the object contains remote
linkEA entry, then it will be recorded in the namespace LFSCK
tracing file.

During the second-stage scanning: if there is no name entry to
reference the MDT-object that has no linkEA also, then insert
the MDT-object into .lustre/lost+found/MDTxxxx/ with the name
${FID}-O-${conflict_version}; if the remote parent is lost,
then re-create the remote parent and re-insert the name entry;
if the parent directory exists, but the name entry corresponding
to the linkEA is lost, then add the missed name entry back to the
namespace (if the linkEA entries count is not more the MDT-object
hard links count).

Simplify the namespace LFSCK tracing file: only record the FIDs
that need to be double scanned, not other flags.

Other code cleanup.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ibc87ae9a5c6b7f67a9215140cf2cb89640bce0a9