LU-4788 lfsck: repair orphan MDT-object
Firstly, for the case that the parent MDT-object stored in some
MDT-object's linkEA entry is lost, the LFSCK will re-create the
parent object as an orphan and insert it into
.lustre/lost+found/MDTxxxx/ directory with the name:
${FID}-P-${conflict_version}.

If there is a linkEA entry that back references a name entry under
some parent directory (that can be a normal directroy, also can be
an orphan parent created as above), but such parent directory does
not has the claimed name entry. On the other hand, the MDT-object's
linkEA entries count is not larger than the MDT-object's hard link
count. Under such case, it is quite possible that the name entry is
lost. Then the LFSCK will add the name entry back to the namespace.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ie9c585f08fa9f371fa3ed1281c9c814606c4a2ea