LU-4788 osd: get PFID from linkEA for remote dir
For a directory which parent resides on remote MDT, to satisfy the
local e2fsck, we insert it into the /REMOTE_PARENT_DIR locally. On
the other hand, to make the lookup(..) on the directroy can return
the real parent FID, we append the real parent FID after its ".."
name entry in the /REMOTE_PARENT_DIR.

Unfortunately, such PFID-in-dirent cannot be preserved via file-level
backup. So after the restore, we cannot get the right parent FID from
its ".." name entry in the /REMOTE_PARENT_DIR. Under such case, since
we have stored the real parent FID in the directory object's linkEA,
we can parse the linkEA for the real parent FID.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Icf1e24ec911818b3a49a253f67c72334a4b75712