LU-4788 lfsck: repair the lost name entry
If there is a linkEA entry that back references a name entry under
some parent directory, but such parent directory does not has the
claimed name entry. On the other hand, the linkEA entries count is
not larger than the MDT-object's hard link count. Under such case,
it is quite possible that the name entry is lost. Means the object
is an orphan. Then the LFSCK should add the name entry back to the
namespace according to the linkEA entry.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ia84f0a545c9dd87976de3d198951077a4c8da671