LU-4788 lfsck: repair invalid nlink count
If the namespace LFSCK has verified all the known name entries during
the first-stage scanning, then the object's linkEA is trustable. So
if the object's nlink attribute does not match the linkEA entries
count, then the LFSCK will repair the nlink attribute according to
the linkEA entries count.

One exception is that: the linkEA space is limited, if there are too
much hard links on the object and exceeds the object's linkEA space
limitation, then some name entry cannot be recorded in the linkEA.
Under such case, we will add some flags in the LFSCK tracing file
for related objects. Then the LFSCK can skip the nlink attribute
verification for those objects during the second-stage scanning.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Iedc676e8cc06a52f55e82372e6dc8b30008e20f4