LU-5304: no need to zero out the allocated memory when GFP_ZERO is set
The kernel will zero out the memory allocated when GFP_ZERO is set,
so no need to zero it out again.

It is possible the original intent was to memset the allocated memory
if GFP_ZERO was not set. However the only two existing callers of this
function already set GFP_ZERO.

Signed-off-by: Frank Zago <fzago@cray.com>
Change-Id: Ia7f365f91aa2ce7a302e4f794706c978222317bb