LU-5516 lfsck: repair the lost name entry
If the MDT-object has a linkEA entry that back references a name
entry under some parent directory, but such parent directory has
no the claimed name entry. On the other hand, if the MDT-object's
linkEA entries count is not larger than the MDT-object's hard links
count. Under such case, it is quite possible that the name entry is
lost. So the object may be an orphan. Then the LFSCK should add the
missed name entry back to the namespace according to the linkEA entry.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ia84f0a545c9dd87976de3d198951077a4c8da671