LU-2740 utils: Add support for --version option
For mount.lustre, lfs, and lctl, a --version option was
added that will print the LUSTRE_VERSION_STRING.

lctl has an option called lustre_build_version that prints
both the Lustre version and the lctl version, which is
BUILD_VERSION. Printing the lctl version was removed when
lustre_build_version is called since
BUILD_VERSION and LUSTRE_VERSION_STRING are very different
strings and the two version strings printed by
lustre_build_version, lctl and Lustre, are the same. So,
the information printed by lustre_build_version is redundant
and would be different from the newly added --version option.

Also, mount.lustre prints the version number when printing
help/usage. This functionality is unique to mount.lustre and
was removed to make all Lustre utilities have a more uniform
output.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I301258562f51640f041dd7ac6aa695450270c385