LU-5515 lfsck: repair bad file type in name entry
If the file type stored in the name entry does not match the file
type claimed by the object, then trust the object and update the
file type in the name entry.

At the same time, if the object does not recognize the name entry
(via the linkEA), then it is quite possible that the name entry is
corrupted. Then remove the bad name entry from namespace.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I05348b70984b6ae8d8a068b63eab38e0273239c0