LU-5341 ptlrpc: rpc times out in unregistering phase
If the reply to an rpc is received before Lustre is notified
that the request has been sent, the task making the request may sleep
until the rpc times out. Both the rq_reply_unlink and rq_req_unlink
flags must be cleared before an rpc can be completed. When the reply
is received, rq_reply_unlink is reset and the  waiter on the rpc wait
queue is woken up. If rq_req_unlink has not been reset at that time,
the rpc is unregistered and the waiter goes back to sleep. When
Lustre is then notified that the request has been sent,
request_out_callback resets rq_req_unlink but it only wakes up the
waiter if the event type is an UNLINK. If the event type
is a SEND, the waiter waits until the timeout on the the wait queue
triggers.

This mod also does a bit of code cleanup: removes an unnecessary
check for a condition that has already been tested in an LASSERT;
adds the request and reply unlink flags to DEBUG_REQ messages.

Signed-off-by: Ann Koehler <amk@cray.com>
Change-Id: I1cbf465d11750c1495a3e8bf2267c392e9c08414