LU-5506 lfsck: skip orphan OST-object handling for failed OSTs
The layout LFSCK will record the failed OSTs in the LFSCK tracing
file (lfsck_layout) during the first-stage scanning, then when moves
to the second-stage scanning, the layout LFSCK will know which OSTs
contain the OST-objects that have not been verified or failed to be
verified during the first-stage scanning. Then the layout LFSCK will
skip the orphan OST-objects handling on those OSTs. But other OSTs
can be handled as normal case without affected by the failed OSTs.

This patch also builds the framework of recording failed MDTs for
namespace LFSCK

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I4d7a9fc2e22cb8c6ef1c4cf73383ec588c95da53