LU-4298 utils: do not create file with no striping info
'lfs setstripe' should not create a file with no striping
information if bad values are given. The input parameter
check in llapi_file_open_pool is moved before the file open
to catch bad parameters before the file is created.

The O_EXCL open flag is added to llapi_file_create_pool
so we don't try and modify stripe parameters of existing
files.

Signed-off-by: Sebastien Buisson <sebastien.buisson@bull.net>
Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I19beef973dc3a62bf2cb443e02195896772702fc