LU-4821 tests: clean up get_mds_num to be get_mdt_idx
The get_mds_num() helper function returned the MDS facet number
instead of the MDT index, but it is actually more useful if it
returns the MDT index, since this can be extracted directly
from a file via "lfs getdirstripe -M".

Rename the function and fix the few callers to use get_mdt_idx().

Fix some coding style issues in the affected functions.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I23d6b4c8da6e50ee1d94cf7be455265bbf3ebbe5