LU-5365 lod: return -ENXIO if an OST pool does not exist
In the current implementation on MDS, an invalid OST pool name
can be accepted while setting striping information on a file
via ioctl() or fsetxattr() interfaces directly. This patch fixes
the codes which handle the return value of lod_find_pool() to
return an error (-ENXIO) if the OST pool does not exist.

Signed-off-by: Jian Yu <jian.yu@intel.com>
Change-Id: I58d20d4e11fc7ade2189bb13542f9a2591f7ccc9