LU-XXXX: obd: don't let coh_page_bufsize grow
lov_init_raid0 is incrementing coh_page_bufsize with the size of each
stripe. However when the layout is resized, it starts with the
previous, already incremented coh_page_bufsize, adding more to
it. Eventually, given several layout resize, coh_page_bufsize reaches
64KiB and overflows back to a small value. At that point the object is
not big enough to store the requested data, and this leads to memory
scribble.

So reset coh_page_bufsize in lov_delete_raid0.

Change-Id: I0f55664cc761bb8e4d0c714132d46f5e9ea76804
Signed-off-by: frank zago <fzago@cray.com>