LU-3534 lod: record update for cross-MDT operation
Packing updates for cross-MDT operation in the buffer, and
the maximum update record size is about 1M, which is enough
to create 2k stripes for now. To save the save, these update
records will use different packing format with OSP RPC,
\see lustre/target/update_records.c

These updates will be stored in all of MDTs(in the later patch).
During the recovery, master MDT will redo the operation according
to these updates records.

Change-Id: Ic1919ab1c3d2eeca9ef027e2309c42201e3f7a74
Signed-off-by: Wang Di <di.wang@intel.com>