LU-4239 mdt: Remove pli_fids & pli_fidcount
On directory paths 100 or more deep, fid2path
gives a spurious:
ioctl err -75: Value too large for defined data type

This bug occurs because all of the fids found during
a path lookup are stored in pli_fids and the
array is too small for very deep paths.

The stored fids are not used except for the last one.
This patch removes the array entirely and stores the
relevant fid in pli_fid instead.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I357b4d91b9c65a836d289d5815c763306ac34565