LU-2182 llapi: Use new layout for lfs setstripe
This patch moves the lfs_setstripe and llapi_file_open
functions to the new layout api. We also expand layout
api to include a verification function of the layout
created by the user. llapi_layout_verify test to see
if the target lustre file system supports what the
user wants.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: Ie2eb1538881d3e6b070c45d50a4a8c3dd87052c7