LU-2827 ldlm: wrong handling of ldlm resends (ORNL version)
If a successful reply is lost for an intent
lock request, MDS will not correctly recover
from this situation on resend. Instead, a new
lock is created.

This patch also adds handling of resend during layout-lock.

This patch is the merge + b2_4 back-port of the following
2 patches to the special ORNL 2.4 server tree. This is done
to see what regression are in the LU-2827 patch that has
been modified so it can be applied to the ORNL lustre code
base :

Lustre-commit: I9a23c13313d5168d0c232fdc0acbdab330089356
Lustre-change: http://review.whamcloud.com/5978/

Lustre-commit: Ideb551779a884d77ded8036c513f62557b2dfcde
Lustre-change: http://review.whamcloud.com/10378

Signed-off-by: Vitaly Fertman <vitaly_fertman@xyratex.com>
Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>
Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: Ie46cb83f01895f07f28496cbdb8a5cc1dc500210