LU-2182 llite: Make setxattr behave like LOV_SETSTRIPE ioctl
While testing the layout enhancements as the back end to the
lustre utilites I encountered some failures which I traced
back to ll_setxattr. The first problem was several tests use
lfs setstripe to change the first OST index to a specific
value. That value was never being set due to ll_setxattr
always setting stripe offset to -1 in the case when stripe
offset is set to zero. The second problem encountered was
any error condition reported by ll_lov_setstripe_ea_info was
being ignored. This is incorrect behavior which makes it seem
lfs setstripe can be called multiple times on a file which
really is not the case.

Change-Id: Idca8e23a9263d5a7fb4523719a2f25b80ce67462
Signed-off-by: James Simmons <uja.ornl@gmail.com>